add tests and resolve problems

+++ models/facility.go
@@ -4,6 +4,6 @@ type Facility struct {
 	Id        string `json:"facility_id"`
 	City      string `json:"city"`
 	Country   string `json:"country"`
-	Latitude  int64  `json:"latitude"`
-	Longitude int64  `json:"longitude"`
+	Latitude  float64  `json:"latitude"`
+	Longitude float64  `json:"longitude"`
 }
